[core & app] pass c2c msg random to app layer

+++ src/core/packet/message/MessageContentHead.ts
@@ -4,7 +4,7 @@ export const MessageContentHead = new NapProtoMsg({
     type: ProtoField(1, ScalarType.UINT32), // message type
     subType: ProtoField(2, ScalarType.UINT32, true), // message subtype (0x211/0x2dc/0x210 and other system message subtypes, same as c2c_cmd)
     c2cCmd: ProtoField(3, ScalarType.UINT32, true), // c2c message subtype
-    random: ProtoField(4, ScalarType.INT64, true),
+    random: ProtoField(4, ScalarType.UINT32, true),
     sequence: ProtoField(5, ScalarType.UINT32, true),
     timestamp: ProtoField(6, ScalarType.INT64, true),
     pkgNum: ProtoField(7, ScalarType.INT64, true), // number of packages, not 1 when the message needs to be sent in packages
